Added MIT License and Disclaimer
A new license file has been added to the project, adopting the MIT License. The copyright notice and permission notice are included in this file. Additionally, a disclaimer has been appended to both the license file and README.md, clarifying that this project is not affiliated with Teamwork.com. Lastly, a minor adjustment was made in README.md to change the heading level of "Using a Configuration File".